chore!: Basically entire bot refactor
