+++ bot/database/models/__init__.py
@@ -1,0 +1,4 @@
+from .base import Base
+from .user import MoodModel, UserModel
+
+__all__ = ["Base", "MoodModel", "UserModel"]
