+++ bot/database/models/base.py
@@ -6,9 +6,9 @@ from typing import Annotated
 from sqlalchemy import BigInteger, text
 from sqlalchemy.orm import DeclarativeBase, mapped_column
 
-int_pk = Annotated[int, mapped_column(primary_key=True, unique=True, autoincrement=False)]
+int_pk = Annotated[int, mapped_column(primary_key=True, unique=True, autoincrement=True)]
 big_int_pk = Annotated[int, mapped_column(primary_key=True, unique=True, autoincrement=False, type_=BigInteger)]
-created_at = Annotated[datetime.datetime, mapped_column(server_default=text("TIMEZONE('utc', now())"))]
+created_at = Annotated[datetime.datetime, mapped_column(server_default=text("current_timestamp"))]
 
 
 class Base(DeclarativeBase):
