+++ bot/instruction_templates/Mistral.yaml
