+++ bot/instruction_templates/Vicuna-v1.1.yaml
