+++ config.yaml
@@ -140,7 +140,19 @@ models:
     bad_russian: false
     price: 0
 
-default_model_id: 16
+  - id: 18
+    name: deepseek/deepseek-chat-v3.1:free
+    template: null
+    bad_russian: false
+    price: 0
+
+  - id: 19
+    name: tngtech/deepseek-r1t-chimera:free
+    template: null
+    bad_russian: false
+    price: 0
+
+default_model_id: 18
 
 paths:
   db: "./db.db"
