--- src/database/models/__init__.py
@@ -1,4 +1,0 @@
-from .base import Base
-from .user import UserModel
-
-__all__ = ["Base", "UserModel"]
